gridnav: split the unit and sqrt portions of the cost for better accuracy. This is a bit slower but whatever, it is easy to remove in place of doubles if I ever need to squeeze an extra fraction of a second out.

+++ gridnav/scenario.hpp
@@ -24,7 +24,7 @@ private:
 	unsigned int w, h;
 	unsigned int x0, y0;
 	unsigned int x1, y1;
-	float opt;
+	double opt;
 };
 
 struct Scenario {
